feat: Limit foreground service duration and add stop service button
This commit limits the duration of the foreground service and adds a button to stop the service to `MainActivity
`.

The `MyForegroundService` now stops itself after 3 seconds by calling `stopSelf()`.
The `MainActivity` now includes a button to stop the `MyForegroundService` by calling `stopService()`.
This change also stops the `MyForegroundService` before starting `MyService` to
 prevent both services from running simultaneously.