feat: Add JobWork
This commit adds the ability to use `JobWorkItem` with a `JobScheduler`.

The `MyJobService` now processes
 work items sequentially using a `while` loop and dequeuing work items with `dequeueWork`.
Each work item includes a page number that is used in the timer loop.
This change also adds the `newIntent` function to create an intent with a page number extra.
The `MainActivity` now enqueues
 a work item with a page number when the button is clicked.
This change also removes the `setPersisted(true)` option from the `JobInfo` builder.