feat: Add WorkManager example
This commit adds an example of using WorkManager.

The `MainActivity` now includes a button to enqueue a `OneTime
WorkRequest`.

This change also adds the `MyWorker` class, which receives the page number as input data and simulates a timer loop.
The worker requires the device to be charging.
This change also adds the `work-runtime-ktx` dependency to the project.